Fix: Refactor codes for Login.js file To improve the structure and apply Styled-components appropriately Ref: #134
