+++ package.json
@@ -11,7 +11,6 @@
     "axios": "^0.26.1",
     "chart.js": "^3.7.1",
     "connected-react-router": "^6.9.2",
-    "date-fns": "^2.28.0",
     "dayjs": "^1.11.0",
     "dom-to-image": "^2.6.0",
     "dotenv": "^16.0.0",
