+++ src/components/Header.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from "react";
 import styled from "styled-components";
 import { useDispatch, useSelector } from "react-redux";
-import Login from "./Login";
+import Login from "./Login/Login";
 import { actionCreators as userActions } from "../redux/modules/user";
 import { history } from "../redux/configureStore";
 import NicknameModal from "./NicknameModal/NicknameModal";
