+++ src/components/Login/Login.js
@@ -1,12 +1,27 @@
 import React from 'react';
 import styled from 'styled-components';
 import { useDispatch } from 'react-redux';
-import { actionCreators as userActions } from '../redux/modules/user';
-import { history } from "../redux/configureStore";
-import { KAKAO_AUTH_URL } from '../shared/social/Kakao';
-import Signup from './Signup';
-import { alertNew } from '../shared/alert';
+import { actionCreators as userActions } from '../../redux/modules/user';
+import { history } from "../../redux/configureStore";
+import { KAKAO_AUTH_URL } from '../../shared/social/Kakao';
+import Signup from '../Signup';
+import { alertNew } from '../../shared/alert';
 import ReactGA from "react-ga";
+import { 
+  ModalWrapper, 
+  ModalBox, 
+  CloseButton, 
+  TyingLogo, 
+  InfoTitle, 
+  LoginButton, 
+  LoginButton2, 
+  KakaoButton, 
+  KakaoButton2,
+  NoSignup,
+  TyingWelcomeLogo,
+  LoginTitle,
+  LoginForm,
+  LoginButton3 } from './style';
 
 
 const Login = (props) => {
@@ -51,7 +66,7 @@ const Login = (props) => {
     <React.Fragment>
       <ModalWrapper>
         <ModalBox>
-          <div className='close-button' onClick={()=>{dispatch(userActions.setLoginModal(false));}}>
+          <CloseButton onClick={()=>{dispatch(userActions.setLoginModal(false));}}>
             <svg
               width='24'
               height='24'
@@ -66,11 +81,11 @@ const Login = (props) => {
                 fill='#D2D2D2'
               />
             </svg>
-          </div>
+          </CloseButton>
 
           {modal_state === 'welcome' && (
             <div>
-              <div className='tying-logo'>
+              <TyingLogo>
                 <svg
                   width='113'
                   height='42'
@@ -100,25 +115,25 @@ const Login = (props) => {
                   />
                   <circle cx='52' cy='5' r='5' fill='#FF2E00' />
                 </svg>
-              </div>
+              </TyingLogo>
 
-              <div className='info-title'>
+              <InfoTitle>
                 <p>로그인하고 모든 기능을 이용해보세요.</p>
                 <p>필요한 시간은 단, 3초!</p>
-              </div>
+              </InfoTitle>
 
-              <div className='login-button'>
-                <button
+              <LoginButton>
+                <LoginButton2
                   onClick={() => {
                     setModalState('login');
                   }}
                 >
                   로그인
-                </button>
-              </div>
+                </LoginButton2>
+              </LoginButton>
 
-              <div className='kakao-button'>
-                <button
+              <KakaoButton>
+                <KakaoButton2
                   onClick={() => {
                     window.location.href = KAKAO_AUTH_URL;
                   }}
@@ -140,10 +155,10 @@ const Login = (props) => {
                     />
                   </svg>
                   카카오 로그인
-                </button>
-              </div>
+                </KakaoButton2>
+              </KakaoButton>
 
-              <div className='no-signup'>
+              <NoSignup>
                 아직 회원이 아닌가요?{' '}
                 <span
                   onClick={() => {
@@ -152,14 +167,14 @@ const Login = (props) => {
                 >
                   회원가입
                 </span>
-              </div>
+              </NoSignup>
             </div>
           )}
 
           {modal_state === 'login' && (
-            <div className='modal-wrapper'>
+            // <ModalWrapper>
               <div className='white_block'>
-                <div className='tying-welcome-logo'>
+                <TyingWelcomeLogo>
                   <svg
                     width='368'
                     height='52'
@@ -173,13 +188,13 @@ const Login = (props) => {
                     />
                     <circle cx='295' cy='15' r='5' fill='#FF2E00' />
                   </svg>
-                </div>
+                </TyingWelcomeLogo>
 
-                <div className='login-title'>
+                <LoginTitle>
                   <span>로그인</span>
-                </div>
+                </LoginTitle>
 
-                <div className='login-form'>
+                <LoginForm>
                   <input
                     type='id'
                     placeholder='아이디를 입력하세요.'
@@ -196,10 +211,10 @@ const Login = (props) => {
                         login();
                     }}}
                     />
-                  <button onClick={login}>로그인</button>
-                </div>
+                  <LoginButton3 onClick={login}>로그인</LoginButton3>
+                </LoginForm>
               </div>
-            </div>
+            // </ModalWrapper>
           )}
 
           {modal_state === 'signup' && <Signup setModalState={setModalState} />}
@@ -208,282 +223,5 @@ const Login = (props) => {
     </React.Fragment>
   );
 };
-
-const ModalWrapper = styled.div`
-  width: 100vw;
-  height: 100vh;
-  position: fixed;
-  top: 0;
-  left: 0;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 10000;
-`;
-
-const ModalBox = styled.div`
-  width: 24.9vw;
-  /* height: 478px; */
-  background-color: #f9f9f9;
-  border-radius: 1.04vw;
-  /* display: flex;
-  justify-content: center; */
-  padding: 3.91vw;
-  box-sizing: border-box;
-  position: relative;
-  transition: 0.5s;
-
-  .close-button {
-    position: absolute;
-    width: 1.25vw;
-    height: 1.25vw;
-    left: 22.45vw;
-    top: 1.09vw;
-    z-index: 10001;
-    cursor: pointer;
-
-    >svg{
-      width: 1.25vw;
-      height: 1.25vw;
-    }
-
-    svg path{
-      transition: 0.3s;
-    }
-
-    &:hover {
-      svg path {
-        fill: #464646;
-      }
-    }
-  }
-
-  .tying-logo {
-    display: block;
-    margin: auto;
-    width: 7.5vw;
-    height: 2.6vw;
-    >svg{
-      width: 5.89vw;
-      height: 2.19vw;
-    }
-  }
-
-  .info-title {
-    justify-content: center;
-    margin: 1.3vw auto;
-
-    font-family: 'Noto Sans KR';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 1.04vw;
-    line-height: 1.41vw;
-    letter-spacing: -0.015em;
-
-    > p {
-      margin: 0;
-      text-align: center;
-    }
-  }
-
-  .login-button {
-    width: 70%;
-    height: 2.6vw;
-    justify-content: center;
-    margin: 2.34vw auto 0;
-    padding: 0.16vw;
-
-    button {
-      display: block;
-      width: 100%;
-      height: 2.5vw;
-
-      font-family: 'Noto Sans KR';
-      font-style: normal;
-      font-weight: 700;
-      font-size: 0.89vw;
-      line-height: 1.3vw;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      text-align: center;
-      letter-spacing: -0.015em;
-
-      color: #ffffff;
-
-      background: #ff2e00;
-      border: 0.16vw solid #ff2e00;
-      box-sizing: border-box;
-      border-radius: 1.04vw;
-      outline: none;
-
-      cursor: pointer;
-      transition: 0.3s;
-
-      &:hover {
-        background-color: #e6e7e8;
-        border: none;
-      }
-    }
-  }
-
-  .kakao-button {
-    width: 70%;
-    height: 2.6vw;
-    justify-content: center;
-    margin: 0 auto;
-    padding-top: 0.21vw;
-
-    button {
-      display: block;
-      width: 100%;
-      height: 2.5vw;
-      position: relative;
-
-      font-family: 'Noto Sans KR';
-      font-style: normal;
-      font-weight: 700;
-      font-size: 0.89vw;
-      /* line-height: 25px; */
-      /* display: flex; */
-      justify-content: center;
-      /* align-items: center; */
-      text-align: center;
-      letter-spacing: -0.015em;
-
-      background: #fee500;
-      box-sizing: border-box;
-      border-radius: 1.04vw;
-      border: none;
-      outline: none;
-
-      cursor: pointer;
-      transition: 0.3s;
-
-      &:hover {
-        background-color: #e6e7e8;
-        border: none;
-      }
-
-      > svg {
-        width: 0.99vw;
-        height: 0.94vw;
-        position: absolute;
-        top: calc(50% - 0.47vw);
-        left: 0.99vw;
-      }
-    }
-  }
-
-  .no-signup {
-    justify-content: center;
-    margin: auto;
-    margin-top: 1.3vw;
-
-    font-family: 'Noto Sans KR';
-    font-style: normal;
-    font-weight: 700;
-    font-size: 0.83vw;
-    text-align: center;
-    letter-spacing: -0.015em;
-    color: #878889;
-
-    span {
-      color: #ff2e00;
-      cursor: pointer;
-    }
-  }
-
-  .tying-welcome-logo {
-    display: flex;
-    
-    >svg{
-      width: 19.17vw;
-      height: 2.71vw;
-    }
-  }
-
-  .login-title {
-    width: 11.61vw;
-    height: 1.82vw;
-    left: 6.61vw;
-    top: 5.1vw;
-    justify-content: center;
-    margin: auto;
-
-    font-family: 'Noto Sans KR';
-    font-style: normal;
-    font-weight: 600;
-    font-size: 0.99vw;
-    line-height: 1.41vw;
-    display: flex;
-    align-items: center;
-    text-align: center;
-    letter-spacing: -0.015em;
-
-    color: #878889;
-  }
-
-  .login-form {
-    margin-top: 2.6vw;
-    text-align: right;
-    input {
-      display: block;
-      margin: auto;
-      width: 70%;
-      height: 2.5vw;
-      font-size: 0.83vw;
-      margin-top: 0.42vw;
-
-      font-family: 'Noto Sans KR';
-      font-style: normal;
-      color: #212529;
-
-      border: 0.05vw solid rgb(222, 226, 230);
-      border-radius: 1.04vw;
-      flex: 1 1 0%;
-      box-sizing: border-box;
-      padding: 0.83vw;
-      background: #fff;
-      outline: none;
-
-      &:focus {
-        border: 0.05vw solid #212529;
-      }
-    }
-    button {
-      display: block;
-      margin: 1.04vw auto 1.25vw;
-      width: 70%;
-      height: 2.5vw;
-
-      font-family: 'Noto Sans KR';
-      font-style: normal;
-      font-weight: 700;
-      font-size: 0.89vw;
-      line-height: 1.3vw;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      text-align: center;
-      letter-spacing: -0.015em;
-
-      color: #ffffff;
-
-      background: #ff2e00;
-      border: 0.16vw solid #ff2e00;
-      box-sizing: border-box;
-      border-radius: 1.04vw;
-      outline: none;
-      transition: 0.3s;
-      cursor: pointer;
-
-      &:hover {
-        background-color: #e6e7e8;
-        border: none;
-      }
-    }
-  }
-`;
 
 export default Login;
